ditch the old typechecker completely
